added sweetalerts

+++ client/src/app/survey-library/survey-management/survey-management.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Survey } from 'src/app/model/survey.model';
 import { SurveyRepository } from 'src/app/model/survey.repository';
+import Swal from 'sweetalert2/dist/sweetalert2.js';
 
 @Component({
   selector: 'app-survey-management',
@@ -25,10 +26,31 @@ export class SurveyManagementComponent implements OnInit {
   onCreateSurvey(): void {
     this.repository.addSurvey(this.newSurvey);
     this.initializeNewSurvey();
+    Swal.fire({
+      title: 'Survey Created',
+      text: 'Select your survey to add questions',
+      icon: 'success'
+    });
   }
 
   onDeleteSurvey(id: string): void {
-    this.repository.deleteSurvey(id);
+
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You will not be able recover this survey.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete forever',
+      cancelButtonText: 'No, I change my mind'
+    }).then((result) => {
+      if (result.value) {
+        this.repository.deleteSurvey(id);
+        Swal.fire({
+          title: 'Survey deleted',
+          icon: 'success'
+        });
+      }
+    });
   }
 
   initializeNewSurvey(): void {
